requirements.txt
tweepy==4.14.0
praw==7.7.1
textblob==0.17.1
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
wordcloud==1.9.2
numpy==1.24.3
plotly==5.15.0
python-dotenv==1.0.0